data_fetch/pyproject.toml
[tool.poetry]
name = "data_fetch"
package-mode = false
version = "0.1.0"
description = ""
authors = ["tanh <tanhaeuser@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
